# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /freelancer:
    get:
      tags:
      - Freelancer
      operationId: getFreelancer
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/FreelancerDto'
    post:
      tags:
      - Freelancer
      operationId: createFreelancer
      parameters:
      - name: registrationRequest
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/FreelancerRegistrationRequest'
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/FreelancerDto'
...
  /auth/switch-role:
    post:
      tags:
      - Auth
      operationId: switchRole
      parameters:
      - name: role
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/RoleType'
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /auth/signup:
    post:
      tags:
      - Auth
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupFormRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/logout:
    post:
      tags:
      - Auth
      operationId: logoutUser
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /auth/login:
    post:
      tags:
      - Auth
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/check-email:
    post:
      tags:
      - Auth
      operationId: checkEmail
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
...
  /auth:
    get:
      tags:
      - Auth
      operationId: getCurrentUser
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/freelancer-registrations:
    get:
      tags:
      - Auth
      operationId: getFreelancerRegistrationRequests
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LightweightRegistrationRequestDto'
  /auth/client-registrations:
    get:
      tags:
      - Auth
      operationId: getClientRegistrationRequests
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LightweightRegistrationRequestDto'
...
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
    AddressDto:
      required:
      - city
      - country
      - houseNumber
      - postalCode
      - region
      - street
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        region:
          type: string
        postalCode:
          type: string
        country:
          type: string
        houseNumber:
          type: string
...
    CategoryType:
      type: string
      enum:
      - WEB_DESIGN
      - WEB_DEVELOPMENT
      - FRONTEND_DEVELOPMENT
      - BACKEND_DEVELOPMENT
      - MOBILE_APP_DEVELOPMENT
      - GAME_DEVELOPMENT
      - WORDPRESS_DEVELOPMENT
      - ECOMMERCE_DEVELOPMENT
      - SOFTWARE_DEVELOPMENT
      - API_DEVELOPMENT
      - GRAPHIC_DESIGN
      - UI_UX_DESIGN
      - LOGO_DESIGN
      - BRAND_IDENTITY_DESIGN
      - VIDEO_EDITING
      - ANIMATION
      - ILLISTRATION
      - THREE_D_MODELING
      - PRESENTATION_DESIGN
      - INTERIOR_DESIGN
      - CONTENT_WRITING
      - COPYWRITING
      - GHOSTWRITING
      - TECHNICAL_WRITING
      - RESUME_WRITING
      - GRANT_WRITING
      - CREATIVE_WRITING
      - SEO_WRITING
      - SCRIPTWRITING
      - EDITING_PROOFREADING
      - TRANSLATION
      - TRANSCRIPTION
      - LANGUAGE_TUTORING
      - SUBTITLING_CAPTIONING
      - LOCALIZATION_SERVICES
      - DIGITAL_MARKETING
      - SOCIAL_MEDIA_MARKETING
      - SEO_SEM
      - EMAIL_MARKETING
      - AFFILIATE_MARKETING
      - INFLUENCER_MARKETING
      - LEAD_GENERATION
      - TELEMARKETING
      - BRAND_STRATEGY
      - CUSTOMER_SUPPORT
      - TECHNICAL_SUPPORT
      - VIRTUAL_ASSISTANT
      - CHAT_SUPPORT
      - HELPDESK_SERVICES
      - DATA_ENTRY
      - WEB_RESEARCH
      - PROJECT_MANAGEMENT
      - CALENDAR_MANAGEMENT
      - DOCUMENT_PREPARATION
      - DATA_ANALYSIS
      - DATA_ENGINEERING
      - MACHINE_LEARNING
      - AI_DEVELOPMENT
      - BUSINESS_INTELLIGENCE
      - DATA_VISUALIZATION
      - BIG_DATA_SERVICES
      - CIVIL_ENGINEERING
      - MECHANICAL_ENGINEERING
      - ELECTRICAL_ENGINEERING
      - STRUCTURAL_ENGINEERING
      - CAD_DRAFTING
      - ARCHITECTURE_DESIGN
      - PRODUCT_DESIGN
      - ACCOUNTING
      - BOOKKEEPING
      - FINANCIAL_ANALYSIS
      - BUSINESS_CONSULTING
      - TAX_PREPARATION
      - HR_CONSULTING
      - FINANCIAL_MODELING
      - LEGAL_CONSULTING
      - CONTRACT_DRAFTING
      - INTELLECTUAL_PROPERTY
      - CORPORATE_LAW
      - FAMILY_LAW_SERVICES
      - COACHING_MENTORING
      - FITNESS_TRAINING
      - MUSIC_PRODUCTION
      - PODCAST_EDITING
      - CAREER_COUNSELING
      - EVENT_PLANNING
      - LIFE_COACHING
      - PLUMBING_SERVICES
      - ELECTRICAL_REPAIR
      - CARPENTRY_SERVICES
      - AUTO_REPAIR
      - HOUSE_CLEANING
      - MOVING_HELP
      - GARDENING_LANDSCAPING
      - CONSTRUCTION_WORK
      - PEST_CONTROL
      - LOCKSMITH_SERVICES
      - PAINTING_SERVICES
      - HVAC_REPAIR
      - APPLIANCE_REPAIR
      - FURNITURE_ASSEMBLY
      - PET_SITTING
      - DOG_WALKING
      - BABYSITTING
      - ELDERLY_CARE
      - SECURITY_SERVICES
      - WINDOW_CLEANING
      - WASTE_REMOVAL
      - RECYCLING_SERVICES
      - CAR_WASHING
      - CAR_SCRAPING
      - BIKE_REPAIR
      - FLOORING_INSTALLATION
      - ROOFING_SERVICES
      - WATER_DAMAGE_RESTORATION
      - POOL_CLEANING
      - SNOW_REMOVAL
      - EVENT_STAFFING
      - CATERING_SERVICES
      - PHOTOGRAPHY_SERVICES
      - MAKEUP_ARTIST
      - TATTOO_ARTIST
      - OTHER
...
    FreelancerRegistrationRequest:
      required:
      - address
      - categories
      - description
      - firstName
      - idDocument
      - lastName
      - phoneNumber
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        phoneNumber:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryType'
        avatarImage:
          type: string
          format: binary
        description:
          type: string
        idDocument:
          type: array
          items:
            type: string
            format: binary
    FreelancerDto:
      required:
      - address
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - rating
      - ratingCount
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          $ref: '#/components/schemas/AddressDto'
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        rating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        avatarImageUrl:
          type: string
...
    RoleType:
      type: string
      enum:
      - ROLE_EMPLOYER
      - ROLE_FREELANCER
      - ROLE_ADMIN
      - ROLE_UNVERIFIED
      - ROLE_CLIENT
    SignupFormRequest:
      required:
      - email
      - password
      - role
      type: object
      properties:
        email:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 32
          minLength: 8
          type: string
        role:
          $ref: '#/components/schemas/RoleType'
    Account:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
        accountType:
          type: string
          enum:
          - ADMIN
          - CLIENT
          - FREELANCER
    AccountStatus:
      type: string
      enum:
      - UNVERIFIED
      - PENDING
      - ACTIVE
      - INACTIVE
      - DELETED
      - BANNED
      - REJECTED
    AuthResponse:
      required:
      - accounts
      - activeRole
      - email
      - roles
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/RoleType'
        activeRole:
          $ref: '#/components/schemas/RoleType'
        avatarImageUrl:
          type: string
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    LightweightRegistrationRequestDto:
      required:
      - accountStatus
      - accountType
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - registrationDate
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
        rejectionReason:
          type: string
        registrationDate:
          type: string
          format: date-time
        accountType:
          type: string
          enum:
          - ADMIN
          - CLIENT
          - FREELANCER
...